adding entries to leaderboard with react is functional
