+++ package.json
@@ -16,12 +16,8 @@
     "bootstrap": "^5.1.1",
     "bootstrap-icons": "^1.5.0",
     "connect-timeout": "^1.9.0",
-    "cookie-session": "^1.4.0",
-    "dotenv": "^10.0.0",
-    "ejs": "^3.1.6",
     "express": "^4.17.1",
     "jquery": "^3.6.0",
-    "mongoose": "^6.0.6",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.13",
     "popper.js": "^1.16.1",
