+++ server.js
@@ -1,5 +1,7 @@
 const express  = require( 'express' ),
     app      = express();
+const bodyParser = require("body-parser");
+const morgan = require("morgan");
 
 const todos = [
     { name:'buy groceries', completed:false }
@@ -11,6 +13,8 @@ const appdata = [
     {'yourname': 'Liam', 'score': 430, 'rank': 3}
 ]
 
+app.use(express.urlencoded({extended: true}));
+app.use(morgan('dev'));
 
 app.use( express.json() )
 
@@ -30,5 +34,87 @@ app.post( '/change', function( req,res ) {
 
     res.sendStatus( 200 )
 })
+
+app.post('/submit', bodyParser.json(), async (req, res) =>{
+    //console.log(req);
+    await addRow(req.body);
+    console.log(appdata);
+    res.json( appdata );
+    /*res.writeHead(200,"OK", {'Content-Type': 'text/plain'});
+    res.end();*/
+})
+function addRow(dataString) {
+    //let jsonApp = JSON.parse(dataString);
+    let jsonApp = dataString;
+    //console.log("jsonApp:\n" + JSON.stringify(jsonApp))
+
+    jsonApp['rank'] = 0;
+    for(let i = 0; i < appdata.length; i++){
+        let user = appdata[i];
+
+        if (jsonApp['yourname'] === user.yourname && (parseInt(jsonApp['score']) > user.score)){
+            deleteRow(i + 1);
+        } else if (jsonApp['yourname'] === user.yourname && (parseInt(jsonApp['score']) <= user.score)) {
+            // return error message
+            return;
+        }
+    }
+    appdata.push(jsonApp);
+    calcRank();
+}
+
+function deleteRow(dataString) {
+    let rankDel = appdata[dataString - 1].rank
+    appdata.splice(dataString - 1, 1);
+    updateRank(rankDel);
+}
+
+function modifyRow(dataString) {
+    let jsonApp = JSON.parse(dataString);
+    for(let user of appdata){
+        if (user.yourname === jsonApp['newName']){
+            return;
+        }
+    }
+    for (let user of appdata){
+        if (user.yourname === jsonApp['oldName']){
+            user.yourname = jsonApp['newName'];
+            return;
+        }
+    }
+}
+
+
+function calcRank(){
+    // for each rank of value rank or lower add 1 to number
+    let newRank = appdata.length;
+    for(let user of appdata){
+        let rank = user.rank
+        if (rank === 0){
+            //Calculate rank value
+            let tempRank = Infinity;
+            for(let otherUser of appdata){
+                if((parseInt(user.score) >= parseInt(otherUser.score)) && (tempRank > otherUser.rank) && (otherUser.rank !== 0)){
+                    tempRank = otherUser.rank;
+                }
+                if(otherUser.rank >= tempRank){
+                    otherUser.rank++;
+                }
+            }
+            if (tempRank !== Infinity){
+                newRank = tempRank;
+            }
+            user.rank = newRank;
+        }
+    }
+} //calculates and updates ranks when users are added
+function updateRank(rankDel){
+    for(let user of appdata){
+        if (user.rank > rankDel){
+            user.rank--;
+        }
+    }
+    calcRank();
+} //updates ranks when users are deleted
 
 app.listen( 8080 )
