+++ src/index.html
@@ -10,11 +10,15 @@
               href="https://fonts.googleapis.com/css?family=ZCOOL+KuaiLe">
         <link rel="stylesheet"
               href="https://fonts.googleapis.com/css?family=ABeeZee">
+        <script type="module" src="index.js"></script>
+        <script type="application/javascript" src="page.js" defer>
+
+        </script>
     </head>
     <body>
     <div class="general gameArea">
         <h1>Flappy Block</h1>
-        <button id="startBtn" onclick="startGame()">Start Game</button>
+        <button id="startBtn" onClick="startGame()">Start Game</button>
         <p>Score: <span id="printScore">0</span></p>
         <br>
         <div>
@@ -39,6 +43,5 @@
             </form>
         </div>
     </div>
-        <script type="module" src="index.js" defer></script>
     </body>
 </html>
